+++ lib/sec/port/libc.h
@@ -33,7 +33,7 @@ typedef
 struct Qid
 {
 	uvlong	path;
-	uint32_t	vers;
+	unsigned long	vers;
 	uchar	type;
 } Qid;
 
@@ -44,9 +44,9 @@ struct Dir {
 	uint	dev;	/* server subtype */
 	/* file data */
 	Qid	qid;	/* unique id from server */
-	uint32_t	mode;	/* permissions */
-	uint32_t	atime;	/* last read time */
-	uint32_t	mtime;	/* last write time */
+	unsigned long	mode;	/* permissions */
+	unsigned long	atime;	/* last read time */
+	unsigned long	mtime;	/* last write time */
 	vlong	length;	/* file length: see <u.h> */
 	char	*name;	/* last element of path */
 	char	*uid;	/* owner name */
@@ -178,8 +178,8 @@ extern	int	_IOUNIT(int);
 
 #define ERRMAX 128
 
-extern	void	setmalloctag(void*, uint32_t);
-extern	uint32_t	getcallerpc(void*);
+extern	void	setmalloctag(void*, unsigned long);
+extern	unsigned long	getcallerpc(void*);
 
 /* Used in libsec.h and not picked up in earlier type definitions */
 typedef unsigned int u32int;
@@ -196,7 +196,7 @@ extern	vlong	nsec(void);
 
 extern void sysfatal(char*, ...);
 
-extern	uint32_t	truerand(void);			/* uses /dev/random */
+extern	unsigned long	truerand(void);			/* uses /dev/random */
 extern	int	getfields(char*, char**, int, int, char*);
 
 #pragma	varargck	type	"lld"	vlong
